Arrange postgres db files
